Make records list

make_records_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make records list. The function returns Records (a list).

make_records_ls(
  ds_tb,
  alternative_var_nm_1L_chr = "alternative",
  card_id_var_nm_1L_chr = "card_id",
  case_choices_mat = matrix(),
  choice_var_nm_1L_chr = "choice",
  choice_vars_pfx_1L_chr = "DCE_B",
  cost_var_nm_1L_chr = "Cost",
  ds_dfidx = NULL,
  flags_ls = list(),
  opt_out_var_nm_1L_chr = "opt_out",
  person_card_uid_var_nm_1L_chr = "PERSONcard_id",
  person_uid_var_nm_1L_chr = "person_uid"
)

Arguments

ds_tb

Dataset (a tibble)

alternative_var_nm_1L_chr

Alternative variable name (a character vector of length one), Default: 'alternative'

card_id_var_nm_1L_chr

Card identity variable name (a character vector of length one), Default: 'card_id'

case_choices_mat

Case choices (a matrix), Default: matrix()

choice_var_nm_1L_chr

Choice variable name (a character vector of length one), Default: 'choice'

choice_vars_pfx_1L_chr

Choice variables prefix (a character vector of length one), Default: 'DCE_B'

cost_var_nm_1L_chr

Cost variable name (a character vector of length one), Default: 'Cost'

ds_dfidx

Dataset (an indexed data frame), Default: NULL

flags_ls

Flags (a list), Default: list()

opt_out_var_nm_1L_chr

Opt out variable name (a character vector of length one), Default: 'opt_out'

person_card_uid_var_nm_1L_chr

Person card unique identifier variable name (a character vector of length one), Default: 'PERSONcard_id'

person_uid_var_nm_1L_chr

Person unique identifier variable name (a character vector of length one), Default: 'person_uid'

Value

Records (a list)