Remove red flag cases

remove_red_flag_cases() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove red flag cases. Function argument data_tb specifies the object to be updated. Argument flags_max_1L_int provides the object to be updated. The function returns Data (a tibble).

remove_red_flag_cases(
  data_tb,
  flags_max_1L_int = 0,
  flags_tot_var_nm_1L_chr = "red_flag_count_int",
  qltv_ax_green_flag_1L_chr = "Green",
  qltv_ax_var_nm_1L_chr = character(0)
)

Arguments

data_tb

Data (a tibble)

flags_max_1L_int

Flags maximum (an integer vector of length one), Default: 0

flags_tot_var_nm_1L_chr

Flags total variable name (a character vector of length one), Default: 'red_flag_count_int'

qltv_ax_green_flag_1L_chr

Qualitative assessment green flag (a character vector of length one), Default: 'Green'

qltv_ax_var_nm_1L_chr

Qualitative assessment variable name (a character vector of length one), Default: character(0)

Value

Data (a tibble)