Add choice models

add_choice_mdls() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add choice models. The function returns Models (a list).

Usage

add_choice_mdls(
  mdls_ls = list(),
  dce_design_ls,
  mdl_params_ls,
  records_ls,
  exclude_chr = character(0),
  indl_predrs_chr = NA_character_,
  include_int = 1:4,
  max_concepts_1L_int = 2L,
  min_threshold_1L_int = 2L,
  nbr_of_clss_1L_int = 2L,
  purpose_chr = "attributes",
  significant_at_1L_dbl = 0.05,
  ...
)

Arguments

mdls_ls

Models (a list), Default: list()

dce_design_ls

Discrete choice experiment design (a list)

mdl_params_ls

Model parameters (a list)

records_ls

Records (a list)

exclude_chr

Exclude (a character vector), Default: character(0)

indl_predrs_chr

Individual predictors (a character vector), Default: 'NA'

include_int

Include (an integer vector), Default: 1:4

max_concepts_1L_int

Maximum concepts (an integer vector of length one), Default: 2

min_threshold_1L_int

Minimum threshold (an integer vector of length one), Default: 2

nbr_of_clss_1L_int

Number of classes (an integer vector of length one), Default: 2

purpose_chr

Purpose (a character vector), Default: 'attributes'

significant_at_1L_dbl

Significant at (a double vector of length one), Default: 0.05

...

Additional arguments

Value

Models (a list)